refactor: 공통 코드 제거 및 AuthTokenService로 통합
